Added error handling

+++ src/main.rs
@@ -34,6 +34,14 @@ fn new_project(ctx: &CommandContext) {
 }
 
 fn scaffold_web(name: &str) {
+	if std::path::Path::new(name).exists() {
+		println!("> A directory with the name '{}' already exists.", name);
+		if !Confirm::new("> Overwrite?").ask() {
+			println!("! Aborted");
+			return;
+		}
+	}
+
     std::fs::create_dir_all(name).unwrap();
     std::fs::write(format!("{}/index.html", name), format!(
 "<!DOCTYPE html>
@@ -81,6 +89,13 @@ body {
 }
 
 fn scaffold_py(name: &str, toml: bool) {
+        if std::path::Path::new(name).exists() {
+                println!("> A directory with the name '{}' already exists.", name);
+                if !Confirm::new("> Overwrite?").ask() {
+			println!("! Aborted");
+                        return;
+                }
+        }
 	std::fs::create_dir_all(name).unwrap();
 	std::fs::write(
 		format!("{}/main.py", name),
@@ -106,6 +121,13 @@ fn scaffold_py(name: &str, toml: bool) {
 }
 
 fn scaffold_rust(name: &str, bin: bool, lib: bool,) {
+    if std::path::Path::new(name).exists() {
+                println!("> A directory with the name '{}' already exists.", name);
+                if !Confirm::new("> Overwrite?").ask() {
+						println!("! Aborted");
+                        return;
+                }
+        }
     std::fs::create_dir_all(format!("{}/src", name)).unwrap();
 	if bin {
     	std::fs::write(format!("{}/src/main.rs", name), "fn main() {\n    println!(\"Hello, world!\");\n}\n").unwrap();
